Extract check logic to an interface and extension methods.
The interface for asserting a collection of HttpRequestMessages can be shared between the current HttpRequestMessageAsserter and the checks that have to be written for NFluent. Therefore extracting all common logic into extension methods and adding an interface to HttpRequestMessageAsserter will make it easier to implement the NFluent Checks.
