+++ src/TestableHttpClient/TestableHttpMessageHandler.cs
@@ -68,52 +68,6 @@ namespace TestableHttpClient
             response = new TimeoutHttpResponseMessage();
         }
 
-        /// <summary>
-        /// Validates that requests have been made, throws an exception when no requests were made.
-        /// </summary>
-        public HttpRequestMessageAsserter ShouldHaveMadeRequests()
-        {
-            return new HttpRequestMessageAsserter(Requests).WithUriPattern("*");
-        }
-
-        /// <summary>
-        /// Validates that requests to a specific uri have been made, throws an exception when no requests were made.
-        /// </summary>
-        /// <param name="pattern">The uri pattern to validate against, the pattern supports *.</param>
-        /// <returns>An <seealso cref="HttpRequestMessageAsserter"/> which can be used for further validations.</returns>
-        public HttpRequestMessageAsserter ShouldHaveMadeRequestsTo(string pattern)
-        {
-            if (pattern == null)
-            {
-                throw new ArgumentNullException(nameof(pattern));
-            }
-
-            return new HttpRequestMessageAsserter(Requests).WithUriPattern(pattern);
-        }
-
-        /// <summary>
-        /// Validates that no requests have been made, throws an exception when requests were made.
-        /// </summary>
-        public void ShouldNotHaveMadeRequests()
-        {
-            _ = new HttpRequestMessageAsserter(Requests, true).WithUriPattern("*");
-        }
-
-        /// <summary>
-        /// Validates that no requests to a specific uri have been made, throws an exception when requests were made.
-        /// </summary>
-        /// <param name="pattern">The uri pattern to validate against, the pattern supports *.</param>
-        /// <returns>An <seealso cref="HttpRequestMessageAsserter"/> which can be used for further validations.</returns>
-        public void ShouldNotHaveMadeRequestsTo(string pattern)
-        {
-            if (pattern == null)
-            {
-                throw new ArgumentNullException(nameof(pattern));
-            }
-
-            _ = new HttpRequestMessageAsserter(Requests, true).WithUriPattern(pattern);
-        }
-
         private class TimeoutHttpResponseMessage : HttpResponseMessage
         {
         }
